refactor: Extract shared utilities from worker to reduce duplication
The worker was duplicating compression and WebDAV upload logic that
already existed in other modules. This refactoring:

1. Add compressVolumeFromDb() to compress-volume.ts
   - Streams files from IndexedDB one at a time to zip
   - Prevents memory issues with large volumes (>1GB)
   - Can be used by both main thread and workers

2. Create webdav-upload.ts shared utility
   - uploadFileWithClient() - uses webdav library client
   - uploadToWebDAV() - standalone function for workers
   - ensureFoldersExist() - shared folder creation
   - Both provider and worker use the same webdav library

3. Update unified-file-worker.ts
   - Remove duplicate compression logic (now uses compressVolumeFromDb)
   - Remove duplicate WebDAV upload functions (now imports from shared util)
   - Worker code reduced by ~260 lines

4. Update webdav-provider.ts
   - Use uploadFileWithClient() from shared utility

Benefits:
- Single source of truth for compression and upload logic
- Easier to maintain and fix bugs in one place
- Worker only handles orchestration, not implementation details
